+++ SIMULATIONS/HD/WD/user_param.h
@@ -27,9 +27,9 @@
 #define MB 2 //barionic mass
 
 int buildModel();
-float  interior_ode(float x, float y[], int i);
-float  taylor_ode(float x, float y[], int i);
+double  interior_ode(float x, double y[], int i);
+double  taylor_ode(float x, double y[], int i);
 
-float eos(float valor, int i);
+double eos(double valor, int i);
 
 float massDensity, polytropicExp, polytropicK;
